feat: add RowSortCommand
